🚚 move login and register to a dedicated directory
