+++ front/src/app/app.routes.ts
@@ -1,6 +1,6 @@
 import { Routes } from '@angular/router';
 import { HomePageComponent } from './home-page/home-page.component';
-import { LoginPageComponent } from './login-page/login-page.component';
+import { LoginPageComponent } from './login-register/login-page/login-page.component';
 import { AuthGuard } from './guard/auth.guard';
 
 export const routes: Routes = [
