+++ front/src/app/login-register/register-page/register-page.component.html
@@ -1,0 +1,1 @@
+<p>register-page works!</p>
