+++ front/src/app/login-register/register-page/register-page.component.ts
@@ -1,0 +1,13 @@
+import { Component } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+@Component({
+  selector: 'app-register-page',
+  standalone: true,
+  imports: [CommonModule],
+  templateUrl: './register-page.component.html',
+  styleUrl: './register-page.component.scss'
+})
+export class RegisterPageComponent {
+
+}
